Add ability to disable animations.
