+++ Sources/SwiftUICharts/BarChart/Models/ChartData/BarChartData.swift
@@ -34,7 +34,7 @@ public final class BarChartData: CTBarChartDataProtocol, ChartConformance {
     @Published public var infoView: InfoViewData<BarChartDataPoint> = InfoViewData()
     @Published public var extraLineData: ExtraLineData!
     
-    @Published public var shouldAnimate: Bool = false
+    @Published public var shouldAnimate: Bool
         
     public var noDataText: Text
     
@@ -56,6 +56,7 @@ public final class BarChartData: CTBarChartDataProtocol, ChartConformance {
     ///   - yAxisLabels: Labels for the Y axis instead of the labels generated from data point values.
     ///   - barStyle: Control for the aesthetic of the bar chart.
     ///   - chartStyle: The style data for the aesthetic of the chart.
+    ///   - shouldAnimate: Whether the chart should be animated.
     ///   - noDataText: Customisable Text to display when where is not enough data to draw the chart.
     public init(
         dataSets: BarDataSet,
@@ -63,6 +64,7 @@ public final class BarChartData: CTBarChartDataProtocol, ChartConformance {
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
         chartStyle: BarChartStyle = BarChartStyle(),
+        shouldAnimate: Bool = true,
         noDataText: Text = Text("No Data")
     ) {
         self.dataSets = dataSets
@@ -70,6 +72,7 @@ public final class BarChartData: CTBarChartDataProtocol, ChartConformance {
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = shouldAnimate
         self.noDataText = noDataText
                 
         self.setupLegends()
@@ -224,7 +227,7 @@ public final class BarChartData: CTBarChartDataProtocol, ChartConformance {
     @available(*, deprecated, message: "Please set use other init instead.")
     public init(
         dataSets: BarDataSet,
-        metadata: ChartMetadata = ChartMetadata(),
+        metadata: ChartMetadata,
         xAxisLabels: [String]? = nil,
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
@@ -237,6 +240,7 @@ public final class BarChartData: CTBarChartDataProtocol, ChartConformance {
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = true
         self.noDataText = noDataText
         
         self.setupLegends()
