+++ Sources/SwiftUICharts/BarChart/Models/ChartData/GroupedBarChartData.swift
@@ -12,7 +12,6 @@ import Combine
  Data model for drawing and styling a Grouped Bar Chart.
  
  The grouping data informs the model as to how the datapoints are linked.
- ```
  */
 @available(macOS 11.0, iOS 14, watchOS 7, tvOS 14, *)
 public final class GroupedBarChartData: CTMultiBarChartDataProtocol, ChartConformance {
@@ -36,8 +35,8 @@ public final class GroupedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
     @Published public var viewData: ChartViewData = ChartViewData()
     @Published public var infoView: InfoViewData<GroupedBarDataPoint> = InfoViewData()
     @Published public var extraLineData: ExtraLineData!
-    
-    @Published public var shouldAnimate: Bool = false
+  
+    @Published public var shouldAnimate: Bool
     
     @Published public var groupSpacing: CGFloat = 0
     @Published public var groups: [GroupingData]
@@ -63,6 +62,7 @@ public final class GroupedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
     ///   - yAxisLabels: Labels for the Y axis instead of the labels generated from data point values.   
     ///   - barStyle: Control for the aesthetic of the bar chart.
     ///   - chartStyle: The style data for the aesthetic of the chart.
+    ///   - shouldAnimate: Whether the chart should be animated.
     ///   - noDataText: Customisable Text to display when where is not enough data to draw the chart.
     public init(
         dataSets: GroupedBarDataSets,
@@ -71,6 +71,7 @@ public final class GroupedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
         chartStyle: BarChartStyle = BarChartStyle(),
+        shouldAnimate: Bool = true,
         noDataText: Text = Text("No Data")
     ) {
         self.dataSets = dataSets
@@ -79,6 +80,7 @@ public final class GroupedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = shouldAnimate
         self.noDataText = noDataText
         
         self.setupLegends()
@@ -261,7 +263,7 @@ public final class GroupedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
     public init(
         dataSets: GroupedBarDataSets,
         groups: [GroupingData],
-        metadata: ChartMetadata = ChartMetadata(),
+        metadata: ChartMetadata,
         xAxisLabels: [String]? = nil,
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
@@ -275,6 +277,7 @@ public final class GroupedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = true
         self.noDataText = noDataText
         
         self.setupLegends()
