+++ Sources/SwiftUICharts/BarChart/Models/ChartData/HorizontalBarChartData.swift
@@ -31,8 +31,8 @@ public final class HorizontalBarChartData: CTHorizontalBarChartDataProtocol, Cha
     @Published public var viewData: ChartViewData = ChartViewData()
     @Published public var infoView: InfoViewData<BarChartDataPoint> = InfoViewData()
     @Published public var extraLineData: ExtraLineData!
-    
-    @Published public var shouldAnimate: Bool = false
+
+    @Published public var shouldAnimate: Bool
         
     public var noDataText: Text
     
@@ -54,6 +54,7 @@ public final class HorizontalBarChartData: CTHorizontalBarChartDataProtocol, Cha
     ///   - yAxisLabels: Labels for the Y axis instead of the labels generated from data point values.
     ///   - barStyle: Control for the aesthetic of the bar chart.
     ///   - chartStyle: The style data for the aesthetic of the chart.
+    ///   - shouldAnimate: Whether the chart should be animated.
     ///   - noDataText: Customisable Text to display when where is not enough data to draw the chart.
     public init(
         dataSets: BarDataSet,
@@ -61,6 +62,7 @@ public final class HorizontalBarChartData: CTHorizontalBarChartDataProtocol, Cha
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
         chartStyle: BarChartStyle = BarChartStyle(),
+        shouldAnimate: Bool = true,
         noDataText: Text = Text("No Data")
     ) {
         self.dataSets = dataSets
@@ -68,6 +70,7 @@ public final class HorizontalBarChartData: CTHorizontalBarChartDataProtocol, Cha
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = shouldAnimate
         self.noDataText = noDataText
         
         self.setupLegends()
@@ -261,7 +264,7 @@ public final class HorizontalBarChartData: CTHorizontalBarChartDataProtocol, Cha
     @available(*, deprecated, message: "Please set use other init instead.")
     public init(
         dataSets: BarDataSet,
-        metadata: ChartMetadata = ChartMetadata(),
+        metadata: ChartMetadata,
         xAxisLabels: [String]? = nil,
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
@@ -274,6 +277,7 @@ public final class HorizontalBarChartData: CTHorizontalBarChartDataProtocol, Cha
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = true
         self.noDataText = noDataText
         
         self.setupLegends()
