+++ Sources/SwiftUICharts/BarChart/Models/ChartData/RangedBarChartData.swift
@@ -33,7 +33,7 @@ public final class RangedBarChartData: CTRangedBarChartDataProtocol, ChartConfor
     @Published public var infoView: InfoViewData<RangedBarDataPoint> = InfoViewData()
     @Published public var extraLineData: ExtraLineData!
     
-    @Published public var shouldAnimate: Bool = false
+    @Published public var shouldAnimate: Bool
         
     public var noDataText: Text
     
@@ -55,6 +55,7 @@ public final class RangedBarChartData: CTRangedBarChartDataProtocol, ChartConfor
     ///   - yAxisLabels: Labels for the Y axis instead of the labels generated from data point values.   
     ///   - barStyle: Control for the aesthetic of the bar chart.
     ///   - chartStyle: The style data for the aesthetic of the chart.
+    ///   - shouldAnimate: Whether the chart should be animated.
     ///   - noDataText: Customisable Text to display when where is not enough data to draw the chart.
     public init(
         dataSets: RangedBarDataSet,
@@ -62,6 +63,7 @@ public final class RangedBarChartData: CTRangedBarChartDataProtocol, ChartConfor
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
         chartStyle: BarChartStyle = BarChartStyle(),
+        shouldAnimate: Bool = true,
         noDataText: Text = Text("No Data")
     ) {
         self.dataSets = dataSets
@@ -69,6 +71,7 @@ public final class RangedBarChartData: CTRangedBarChartDataProtocol, ChartConfor
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = shouldAnimate
         self.noDataText = noDataText
         
         self.setupLegends()
@@ -233,7 +236,7 @@ public final class RangedBarChartData: CTRangedBarChartDataProtocol, ChartConfor
     @available(*, deprecated, message: "Please set use other init instead.")
     public init(
         dataSets: RangedBarDataSet,
-        metadata: ChartMetadata = ChartMetadata(),
+        metadata: ChartMetadata,
         xAxisLabels: [String]? = nil,
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
@@ -246,6 +249,7 @@ public final class RangedBarChartData: CTRangedBarChartDataProtocol, ChartConfor
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = true
         self.noDataText = noDataText
         
         self.setupLegends()
