+++ Sources/SwiftUICharts/BarChart/Models/ChartData/StackedBarChartData.swift
@@ -36,7 +36,7 @@ public final class StackedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
     @Published public var infoView: InfoViewData<StackedBarDataPoint> = InfoViewData()
     @Published public var extraLineData: ExtraLineData!
     
-    @Published public var shouldAnimate: Bool = false
+    @Published public var shouldAnimate: Bool
     
     @Published public var groups: [GroupingData]
     
@@ -61,6 +61,7 @@ public final class StackedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
     ///   - yAxisLabels: Labels for the Y axis instead of the labels generated from data point values.   
     ///   - barStyle: Control for the aesthetic of the bar chart.
     ///   - chartStyle: The style data for the aesthetic of the chart.
+    ///   - shouldAnimate: Whether the chart should be animated.
     ///   - noDataText: Customisable Text to display when where is not enough data to draw the chart.
     public init(
         dataSets: StackedBarDataSets,
@@ -69,6 +70,7 @@ public final class StackedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
         chartStyle: BarChartStyle = BarChartStyle(),
+        shouldAnimate: Bool = true,
         noDataText: Text = Text("No Data")
     ) {
         self.dataSets = dataSets
@@ -77,6 +79,7 @@ public final class StackedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = shouldAnimate
         self.noDataText = noDataText
         
         self.setupLegends()
@@ -274,7 +277,7 @@ public final class StackedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
     public init(
         dataSets: StackedBarDataSets,
         groups: [GroupingData],
-        metadata: ChartMetadata = ChartMetadata(),
+        metadata: ChartMetadata,
         xAxisLabels: [String]? = nil,
         yAxisLabels: [String]? = nil,
         barStyle: BarStyle = BarStyle(),
@@ -288,6 +291,7 @@ public final class StackedBarChartData: CTMultiBarChartDataProtocol, ChartConfor
         self.yAxisLabels = yAxisLabels
         self.barStyle = barStyle
         self.chartStyle = chartStyle
+        self.shouldAnimate = true
         self.noDataText = noDataText
         
         self.setupLegends()
