+++ Sources/SwiftUICharts/LineChart/Models/ChartData/LineChartData.swift
@@ -34,7 +34,7 @@ public final class LineChartData: CTLineChartDataProtocol, ChartConformance {
     @Published public var infoView: InfoViewData<LineChartDataPoint> = InfoViewData()
     @Published public var extraLineData: ExtraLineData!
     
-    @Published public var shouldAnimate: Bool = false
+    @Published public var shouldAnimate: Bool
     
     public var noDataText: Text
 
@@ -57,18 +57,21 @@ public final class LineChartData: CTLineChartDataProtocol, ChartConformance {
     ///   - xAxisLabels: Labels for the X axis instead of the labels in the data points.
     ///   - yAxisLabels: Labels for the Y axis instead of the labels generated from data point values.
     ///   - chartStyle: The style data for the aesthetic of the chart.
+    ///   - shouldAnimate: Whether the chart should be animated.
     ///   - noDataText: Customisable Text to display when where is not enough data to draw the chart.
     public init(
         dataSets: LineDataSet,
         xAxisLabels: [String]? = nil,
         yAxisLabels: [String]? = nil,
         chartStyle: LineChartStyle = LineChartStyle(),
+        shouldAnimate: Bool = true,
         noDataText: Text = Text("No Data")
     ) {
         self.dataSets = dataSets
         self.xAxisLabels = xAxisLabels
         self.yAxisLabels = yAxisLabels
         self.chartStyle = chartStyle
+        self.shouldAnimate = shouldAnimate
         self.noDataText = noDataText
 
         self.setupLegends()
@@ -162,7 +165,8 @@ public final class LineChartData: CTLineChartDataProtocol, ChartConformance {
     
     // MARK: Points
     public func getPointMarker() -> some View {
-        PointsSubView(dataSets: dataSets,
+        PointsSubView(chartData: self,
+                      dataSets: dataSets,
                       minValue: self.minValue,
                       range: self.range,
                       animation: self.chartStyle.globalAnimation,
@@ -242,7 +246,7 @@ public final class LineChartData: CTLineChartDataProtocol, ChartConformance {
     @available(*, deprecated, message: "Please set use other init instead.")
     public init(
         dataSets: LineDataSet,
-        metadata: ChartMetadata = ChartMetadata(),
+        metadata: ChartMetadata,
         xAxisLabels: [String]? = nil,
         yAxisLabels: [String]? = nil,
         chartStyle: LineChartStyle = LineChartStyle(),
@@ -253,6 +257,7 @@ public final class LineChartData: CTLineChartDataProtocol, ChartConformance {
         self.xAxisLabels = xAxisLabels
         self.yAxisLabels = yAxisLabels
         self.chartStyle = chartStyle
+        self.shouldAnimate = true
         self.noDataText = noDataText
 
         self.setupLegends()
