+++ Sources/SwiftUICharts/LineChart/Views/MultiLineChart.swift
@@ -50,15 +50,16 @@ public struct MultiLineChart<ChartData>: View where ChartData: MultiLineChartDat
     private let minValue: Double
     private let range: Double
     
+    @State private var startAnimation: Bool
+    
     /// Initialises a multi-line, line chart.
     /// - Parameter chartData: Must be MultiLineChartData model.
     public init(chartData: ChartData) {
         self.chartData = chartData
         self.minValue = chartData.minValue
         self.range = chartData.range
+        self._startAnimation = State<Bool>(initialValue: chartData.shouldAnimate ? false : true)
     }
-    
-    @State private var startAnimation: Bool = false
     
     public var body: some View {
         GeometryReader { geo in
